Написал сбор метрик с пользователя, добавил метрики, собрал контейнеры, разбил агентов для практик, протестировал генерацию изображений, реализовал работу с cdn

+++ backend/src/courses/application/services/lesson.py
@@ -11,23 +11,37 @@ from ...application.dtos import (
     EditLessonSchema,
     LessonSchema,
 )
-from ...domain.entities import AnyContentBlock, Course, Lesson
-from ..repos import (
-    LessonRepository,
-)
+from ...domain.entities import AnyContentBlock, Lesson
+from ..repos import LessonRepository, ModuleRepository
 from .base_course import BaseCourseService
 
 
 class LessonService(BaseCourseService):
-    def __init__(self, repo: LessonRepository, session: AsyncSession) -> None:
+    def __init__(
+        self,
+        lesson_repo: LessonRepository,
+        module_repo: ModuleRepository,
+        session: AsyncSession,
+    ) -> None:
         """Инициализирует объект и сохраняет зависимости, необходимые для дальнейшей работы."""
-        super().__init__(repo=repo, session=session)
+        super().__init__(repo=lesson_repo, session=session)
+        self.module_repo = module_repo
 
-    async def create(self, module_id: UUID, schema: LessonSchema) -> Course:
+    async def create(self, module_id: UUID | None, schema: LessonSchema) -> Lesson:
         """Создаёт урок и инкапсулирует правила этой операции."""
         lesson = await self.repo.create(Lesson(module_id=module_id, **schema.model_dump()))
         await self.session.commit()
         return lesson
+
+    async def assign_module(self, lesson_id: UUID, module_id: UUID) -> None:
+        module = await self.module_repo.exists(module_id)
+        if not module:
+            raise NotFoundError(f"Module with id {module_id} not found")
+        lesson = await self.repo.exists(lesson_id)
+        if not lesson:
+            raise NotFoundError(f"Lesson with id {lesson_id} not found")
+        await self.repo.assign_module(lesson_id, module_id)
+        await self.session.commit()
 
     async def read_content_blocks(self, lesson_id: UUID) -> list[AnyContentBlock]:
         lesson_exists = await self.repo.exists(lesson_id)
@@ -60,7 +74,10 @@ class LessonService(BaseCourseService):
         await self.session.commit()
         return lesson
 
-    async def delete_lesson(self, lesson_id: UUID) -> None:
+    async def delete(self, lesson_id: UUID) -> None:
         """Выполняет действие `delete_lesson`, чтобы поддержать основной сценарий модуля."""
+        lesson_exists = await self.repo.exists(lesson_id)
+        if not lesson_exists:
+            raise NotFoundError(f"Lesson with id {lesson_id} not found")
         await self.repo.delete(lesson_id)
         await self.session.commit()
